// backend/src/types.ts
export interface IUser {
  _id: Types.ObjectId;
  email: string;
  password: string;
  name: string;
  authProvider?: "local" | "google";
  googleId?: string;
  avatar?: string;
  refreshToken?: string;
  apiKey?: string;
  apiKeyExpiresAt?: Date;
  isDeleted: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface IUserDocument extends IUser, Document {
  _id: Types.ObjectId;
  isPasswordCorrect(password: string): Promise<boolean>;
  generateAccessToken(): Promise<string>;
  generateRefreshToken(): Promise<string>;
  generateApiKey(expiryDays?: number): string;
}

export interface IUserModel extends Model<IUserDocument> { }